+++ srbuild/project/project.py
@@ -1,9 +1,9 @@
+from srbuild.project.file_manager import FileManager
 from srbuild.tools.flags import BuildFlags
 from srbuild.graph.graph import Graph
 from srbuild.logger import G_LOGGER
 from typing import Set
 import inspect
-import glob
 import os
 
 global DEFAULT_GRAPH
@@ -25,20 +25,7 @@ class Project(object):
         G_LOGGER.debug(f"Using Root: {self.root_dir}, Build: {self.build}, Dirs: {self.dirs}")
         # Keep track of all files present in project dirs. Since dirs is a set,
         # files is guaranteed to contain no duplicates as well.
-        self.files = []
-        for dir in self.dirs:
-            for path in glob.iglob(os.path.join(dir, "**"), recursive=True):
-                print (path)
-                if os.path.isfile(path):
-                    self.files.append(os.path.abspath(path))
-        # self.files = list(map(os.path.abspath, self.files))
-        G_LOGGER.debug(f"Found {len(self.files)} files")
-        G_LOGGER.verbose(f"{self.files}")
-
-    # TODO: Docstrings
-    # Finds filename in self.files.
-    def find(self, filename):
-        return [path for path in self.files if path.endswith(filename)]
+        self.files = FileManager(self.dirs)
 
     def _target_impl(self, name, sources, flags, compiler, linker, include_dirs, lib_dirs):
         pass