+++ tests/minimal_project/include/factorial.hpp
@@ -1,4 +1,4 @@
 #pragma once
-#include "utils.hpp"
+#include "minimal_project/include/utils.hpp"
 
 int factorial(int n);