+++ tests/minimal_project/include/fibonacci.hpp
@@ -1,4 +1,4 @@
 #pragma once
-#include "utils.hpp"
+#include "include/utils.hpp"
 
 int fibonacci(int n);